Reorganize gnw module
